requirements.txt
certifi
click
dnspython
email-validator
Flask
Flask-Login
Flask-WTF
idna
itsdangerous
Jinja2
MarkupSafe
pip
psycopg2
python-dotenv
setuptools
Werkzeug
wheel
wincertstore
WTForms

